[BPK-1388] Refactor button links
