+++ native/packages/react-native-bpk-component-button-link/src/BpkButtonLink-test.common.js
@@ -16,16 +16,20 @@
  * limitations under the License.
  */
 
+/* @flow */
+
 import { Text } from 'react-native';
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import BpkButtonLink, { propTypes } from './BpkButtonLink';
+import BpkButtonLink from './BpkButtonLink';
 import { ICON_ALIGNMENTS } from './common-types';
 
 const onPressFn = jest.fn();
+
 const commonTests = () => {
   jest.mock('Text', () => 'Text');
+
   describe('BpkButtonLink', () => {
     it('should render correctly', () => {
       const tree = renderer
@@ -34,34 +38,7 @@ const commonTests = () => {
       expect(tree).toMatchSnapshot();
     });
 
-    it('should support the "large" property', () => {
-      const tree = renderer
-        .create(<BpkButtonLink large title="Lorem ipsum" onPress={onPressFn} />)
-        .toJSON();
-      expect(tree).toMatchSnapshot();
-    });
-    it('should support the "large" and secondary property', () => {
-      const tree = renderer
-        .create(<BpkButtonLink large title="Lorem ipsum" onPress={onPressFn} />)
-        .toJSON();
-      expect(tree).toMatchSnapshot();
-    });
-
-    it('should support the "icon" and "large" property', () => {
-      const tree = renderer
-        .create(
-          <BpkButtonLink
-            icon="baggage"
-            large
-            title="Lorem ipsum"
-            onPress={onPressFn}
-          />,
-        )
-        .toJSON();
-      expect(tree).toMatchSnapshot();
-    });
-
-    it('should support having an icon as well as a title', () => {
+    it('should support the "icon" property', () => {
       const tree = renderer
         .create(
           <BpkButtonLink
@@ -115,51 +92,6 @@ const commonTests = () => {
           .toJSON();
         expect(tree).toMatchSnapshot();
       });
-    });
-
-    it('should accept elements as the icon property', () => {
-      expect(
-        propTypes.icon(
-          {
-            iconOnly: true,
-            icon: <Text>foo</Text>,
-          },
-          'icon',
-          'BpkButtonLink',
-        ),
-      ).toBeFalsy();
-    });
-
-    it('should reject theme property when some theme attributes are omitted', () => {
-      // eslint-disable-line max-len
-      expect(
-        propTypes
-          .theme(
-            {
-              theme: {},
-            },
-            'theme',
-            'BpkButtonLink',
-          )
-          .toString(),
-      ).toEqual(
-        'Error: Invalid prop `theme` supplied to `BpkButtonLink`. For button link, the `theme` prop must include `buttonLinkTextColor`',
-      ); // eslint-disable-line max-len
-    });
-
-    it('should accept theme property when correct attributes are supplied', () => {
-      // eslint-disable-line max-len
-      expect(
-        propTypes.theme(
-          {
-            theme: {
-              buttonLinkTextColor: 'blue',
-            },
-          },
-          'theme',
-          'BpkButtonLink',
-        ),
-      ).toBeFalsy();
     });
   });
 };
